Move CLI implementation to dedicated module
- Shorten __main__ to only execute the proper endpoint.
